Un-revert zeppelin's Zip openables, now that the shadowing has a fix
`ZipOpenable` and `ZipDataOpenable` become `Zip.Openable` and `Zip.DataOpenable`.
The first pass reverted these because they shadow the `Openable` they extend; as
with stratiform's, qualifying the base as `extends aperture.Openable` resolves it.

Two names leave the umbrella's exported surface, and the "tried and reverted"
list loses an entry that was never really blocked.
